add starttls option for NotifyEmail
some email providers (like fastmail) do not support STARTTLS, so
(i think) python's smtplib.SMTPSSL must be used instead of smtplib.SMTP for
the socket. also, if SMTPSSL is being used, then startls will not be
called.

if you think this is correct approach to add this setting, I can
continue and add to the docs, and maybe add fastmail to the
WEBBASE_LOOKUP_TABLE.

but perhaps a third protocol would be a better option for SSL without
STARTTLS (other than mailto/mailtos)