Add Twilio API key support
Twilio recently introduced API keys
- blog post: https://www.twilio.com/blog/better-twilio-authentication-csharp-twilio-api-keys
- docs: https://www.twilio.com/docs/usage/api#authenticate-with-http

Using the API key with the REST API is very similar to the previous
flow, however the API key (and its associated secret) are now used
for authentication. The account SID must still be referenced in the
URL.

This commit also introduces a fixup for the "To" field. Prior to this
commit, the "To" field contained a dictionary returned by is_phone_no(),
rather than a properly formatted phone number.