Telegram: enhance bot owner detection and avoid exception
The /getupdates verb on Telegram may return a result where the first entry is a channel post (so no 'message' key: this triggers an exception on line 489).

This commit makes the function parse all results in the response array and returns the bot owner when an entry with 'message' and 'from' keys is found. Generally, this should mean skipping the first result.