Add email related content support (RFC2387).
MIME multipart contents can access other parts when using "multipart/related"
Content-Type.
This allows, for example, the HTML body to embed attached image files accepted
by email clients and email web clients following RFC 2387.
Additionally it is necessary to add the Content-ID header to the attachment
part so the html body can link it.

This commit inserts the Content-ID to each attachment with filename as the tag
value, allowing the HTML body to reference it using:
<img src="cid:FILE_NAME.EXT">

Tested with:
- Outlook
- Thunderbird - Windows and Linux
- IOS Native email app
- GMAIL - Web Interface
- GMAIL - Android and IOS APP